Factored code from command client into rpc client
